When downloading packages, remember where it comes from.
When the same packages is available from different medias (main/release
and main/update for example), urpmi can forget from which media it
downloaded the package,  and this causes issues with key checking. This
commit makes sure we are using the right package id for the package we
downloaded (the same package will have different IDs if it's available
from different medias).