Mark packages as "manually installed"
With the orphan mechanism, there is no way to manually mark a
package that was installed as a dep of another package as not
being a potential orphan. Mimic apt and if an already package
is passed as an argument to urpmi, remove it from the potential
orphans list. Fixes bug #45054.