Save media.cfg files locally to map urpmi.cfg media to media.cfg
To make it possible to associate more metainformation with the media
stored in urpmi.cfg, we need to be able to associate a given urpmi.cfg
medium with the media.cfg files it was created from. To do that, we save
media.cfg files locally in /etc/urpmi/mediacfg.d/$name with an
additional 'url' or 'mirrorlist' files. Those files are used to resolve
the relative names stored in media.cfg so that we can associate them
with the absolute names stored in urpmi.cfg. The mediacfg.d dir is
cleaned up when a medium is removed (if necessary).