be careful when an obsoleted package is listed in potential orphans
When package B obsoletes package A and package A is listed as a
potential orphan (ie it appears in installed-through-deps.list), then
B is marked as a potential orphan too (since "obsoletes" more or less
means that A was renamed to B). However, if B was originally installed
and does not appear in installed-through-deps, that means it's not a
potential orphan and thus it shouldn't be added to the potential orphans
list. Fixes #54590.