reload mirror list cache if it's in an "old" format
old mirror list caches didn't store the time of the product.id file.
When such a file is encountered, make sure we invalidate the cache.
Otherwise, when upgrading from a distro where the mirror list cache
didn't have the product.id mtime to a distro where the mirror list cache
would have it, the mirror list cache won't be invalidated