remember which media local packages came from
It's sometimes possible that a package is available from both a
local medium (CDROM) and a distant one. This causes problems
when the package isn't signed with the same key (this is bad
but happens on MES5 install CDs) as urpmi tends to get confused
about where a package comes from and then check the package against
the wrong key.

Packages are assigned IDs by urpmi, and these IDs are
used to figure out the package media. This ID is also used as an
identifier to "recognize" identical packages coming from different
media, so we need to be careful when manipulating it. In particular,
we can't assign the right ID to every package right away because then
this would break package downloading (we'd download packages available
on local media).

When processing packages to download, if we encounter a local
package, we assign it the right ID. This assumes the list of media
in the blist are sorted by order of relevance (ie prioritary ones
first).