library: ensure priority updates are installed in a single transaction.
This fix is needed in cases where rpm's libdb changes. If the priority updates
happen to be installed in more than one transaction, the %post script in rpm
which deletes the old index files (/var/lib/rpm/__db.*) may happen at the
end of the first of those transactions. This causes the indexes to be
regenerated by the *old* db version and causes the second transaction to fail.

By ensuring the priority updates are handled in a single transaction, the
indexes will be recreated after the urpmi restart and will thus use the
correct, new libdb to create the indexes.

Patch and fix from Thierry Vignaud