Group options by purpose and add sub-headlines for each group.
The urpmi tools share a lot of options like download options, media
selections and so on. Other option groups are package selection,
output options (for urpmf, urpmq) and database options. Grouping them
makes it easier for the user to spot what she needs.

Additionally tool-specific groups will be put near the top of the
options section, while common groups will be put near the end of that
list, followed by rarely used stuff like debugging. This makes it easier
for the user to spot the tool-specific parts.

Grouping may not be perfect and some option may be in a group not
matching 100 percent. But anyway this should be a fairly good state.
followed