Prefer http mirrors when using a mirrorlist
Sort http mirrors to the beginning of the mirror list. This avoids an
rsync mirror being selected as the mirror sometimes, which causes the
preferred aria2 downloader from not being used.

This also has the effect that in practice aria2 will only be given http
mirrors. If this causes issues, the sorting code can be changed to
reverse sort with the presence of "rsync://" instead.